+++ CHANGELOG.md
@@ -11,9 +11,14 @@
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://www.conventionalcommits.org/) specification when making commits.
 
 ---
+## Unreleased
+
+### Features
+
+* add `oapi call -o/--output` for streaming raw response bodies to files
+
 ### [v0.2.1](https://github.com/sichang824/openapi/compare/v0.2.0...v0.2.1) (2026-07-13)
 
 ### Commits
 
 * fix(spec): support OpenAPI 3.1 type unions ([34b9530](https://github.com/sichang824/openapi/commit/34b95306c044f833311062ac022934365fc34be0))
-